backend/requirements.txt
#################  Django and dependencies  ##########################

asgiref==3.6.0
Django>=3.2.4,<3.3
#Django==4.1.7
sqlparse==0.4.3
django-cors-headers==3.13.0
django-filter==22.1
django-mptt==0.14.0

# Storages
django-storages
boto3

# caching
django-redis

# This module is used to generate AutoSlugField and customs extensions
#django-extensions==3.2.1
django-extensions

django-js-asset==2.0.0
qrcode==7.4.2

gunicorn==20.1.0
Pillow==9.2.0
psycopg2>=2.8.6,<2.9

####################### GRAPHQL API and Authentications ###################

graphene==3.2.2
graphene-django==3.0.0
#graphql-core==3.1.0
graphql-core>=3
graphql-relay==3.1.5
#django-graphql-auth==0.3.16
django-graphql-jwt>=0.3.1,<=0.3.4











